Add requirements.nodeps.txt and fix bfloat16 for TabPFN loader
- Install tabpfn<6.0.0 with --no-deps to avoid pandas downgrade that
  causes binary/Python mismatch in the running test environment
- Convert input tensor to float32 before calling numpy() since numpy
  does not support bfloat16

+++ tabpfn/pytorch/loader.py
@@ -52,7 +52,7 @@ class TabPFNWrapper(nn.Module):
         Returns:
             Tensor: Class probability predictions (n_test_samples, n_classes).
         """
-        probas = self.classifier.predict_proba(X_test.numpy())
+        probas = self.classifier.predict_proba(X_test.float().numpy())
         return torch.tensor(probas, dtype=X_test.dtype)
